LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-ignored.tentative-expected.txt


FAIL Changing animation-timeline changes the timeline (sanity check) assert_equals: expected "140px" but got "0px"
FAIL animation-timeline ignored after setting timeline with JS (ScrollTimeline from JS) assert_equals: expected "180px" but got "0px"
FAIL animation-timeline ignored after setting timeline with JS (ScrollTimeline from CSS) assert_equals: expected "140px" but got "0px"
FAIL animation-timeline ignored after setting timeline with JS (document timeline) assert_equals: expected "120px" but got "100.001999px"
FAIL animation-timeline ignored after setting timeline with JS (null) assert_equals: expected "120px" but got "0px"
